+++ geotrek/trekking/parsers.py
@@ -286,6 +286,8 @@ class GeotrekTrekParser(GeotrekParser):
                         # we have to retrieve the missing ones first
                         self.fetch_missing_categories_for_tour_steps(child_trek)
                         self.parse_row(child_trek)
+                        if not self.obj.pk:
+                            continue
                         trek_child_instance = self.obj
                         OrderedTrekChild.objects.update_or_create(
                             parent=trek_parent_instance[0],