+++ geotrek/common/tests/test_parsers.py
@@ -2061,6 +2061,8 @@ class GeotrekAggregatorParserTest(GeotrekParserTestMixin, TestCase):
             string_parser,
         )
         self.assertEqual(Trek.objects.count(), 6)
+        trek = Trek.objects.get(name="Boucle du Pic des Trois Seigneurs")
+        self.assertIn("Lac et glacier", [theme.label for theme in trek.themes.all()])
         self.assertEqual(POI.objects.count(), 2)
         self.assertEqual(1, Trek.objects.get(name="Foo").information_desks.count())
         self.assertEqual(