+++ geotrek/maintenance/templates/maintenance/sql/post_20_views.sql
@@ -37,7 +37,7 @@ SELECT a.id,
        accessmean.label AS "Access",
        a.date_insert AS "Insertion date",
        a.date_update AS "Update date",
-       l.decoupled AS "Decoupled geometry",
+       l.coupled AS "Network-coupled",
        a.geom_3d
 FROM maintenance_intervention a
 LEFT JOIN