+++ docs/changelog.rst
@@ -9,6 +9,9 @@ CHANGELOG
 
 * Ubuntu focal 20.04 debian package is not supported anymore. Please update or migrate your server to Ubuntu 24.04.
 
+**Documentation**
+
+* Add warning: do not use pgRouting 4.x (refs #5049)
 
 2.122.1         (2026-01-08)
 ----------------------------