+++ docs/changelog.rst
@@ -10,6 +10,7 @@ CHANGELOG
 * Fix fetching published steps of parent treks with aggregator parser
 * Allows to add/edit trek despite missing service type pictogram
 * Trek aggregator now handles when child trek is not saved due to errors
+* Aggregator parser now fetches all category pages
 
 
 2.123.0         (2026-01-15)